move gene cnv frequencies; analysis API docs
