+++ malariagen_data/anopheles.py
@@ -1,5 +1,6 @@
 import warnings
 from abc import abstractmethod
+from bisect import bisect_left, bisect_right
 from collections import Counter
 from itertools import cycle
 from typing import Any, Callable, Dict, List, Mapping, Optional, Tuple, Union
@@ -9,6 +10,7 @@ import bokeh.layouts
 import bokeh.models
 import bokeh.palettes
 import bokeh.plotting
+import dask
 import dask.array as da
 import igv_notebook
 import numba
@@ -55,13 +57,18 @@ from .util import (
     jackknife_ci,
     jitter,
     locate_region,
+    parse_multi_region,
     parse_single_region,
     plotly_discrete_legend,
+    region_str,
+    simple_xarray_concat,
 )
 
 AA_CHANGE_QUERY = (
     "effect in ['NON_SYNONYMOUS_CODING', 'START_LOST', 'STOP_LOST', 'STOP_GAINED']"
 )
+DEFAULT_MAX_COVERAGE_VARIANCE = 0.2
+
 
 # N.B., we are in the process of breaking up the AnophelesDataResource
 # class into multiple parent classes like AnophelesGenomeSequenceData
@@ -2028,6 +2035,611 @@ class AnophelesDataResource(
         ds_aa_frq.attrs["title"] = title
 
         return ds_aa_frq
+
+    def gene_cnv(
+        self,
+        region: base_params.regions,
+        sample_sets=None,
+        sample_query=None,
+        max_coverage_variance=DEFAULT_MAX_COVERAGE_VARIANCE,
+    ):
+        """Compute modal copy number by gene, from HMM data.
+
+        Parameters
+        ----------
+        region: str or list of str or Region or list of Region
+            Chromosome arm (e.g., "2L"), gene name (e.g., "AGAP007280"), genomic
+            region defined with coordinates (e.g., "2L:44989425-44998059") or a
+            named tuple with genomic location `Region(contig, start, end)`.
+            Multiple values can be provided as a list, in which case data will
+            be concatenated, e.g., ["3R", "3L"].
+        sample_sets : str or list of str
+            Can be a sample set identifier (e.g., "AG1000G-AO") or a list of
+            sample set identifiers (e.g., ["AG1000G-BF-A", "AG1000G-BF-B"]) or
+            a release identifier (e.g., "3.0") or a list of release identifiers.
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        max_coverage_variance : float, optional
+            Remove samples if coverage variance exceeds this value.
+
+        Returns
+        -------
+        ds : xarray.Dataset
+            A dataset of modal copy number per gene and associated data.
+
+        """
+
+        regions: List[Region] = parse_multi_region(self, region)
+        del region
+
+        ds = simple_xarray_concat(
+            [
+                self._gene_cnv(
+                    region=r,
+                    sample_sets=sample_sets,
+                    sample_query=sample_query,
+                    max_coverage_variance=max_coverage_variance,
+                )
+                for r in regions
+            ],
+            dim="genes",
+        )
+
+        return ds
+
+    def _gene_cnv(self, *, region, sample_sets, sample_query, max_coverage_variance):
+        debug = self._log.debug
+
+        debug("sanity check")
+        assert isinstance(region, Region)
+
+        debug("access HMM data")
+        ds_hmm = self.cnv_hmm(
+            region=region.contig,
+            sample_sets=sample_sets,
+            sample_query=sample_query,
+            max_coverage_variance=max_coverage_variance,
+        )
+        pos = ds_hmm["variant_position"].data
+        end = ds_hmm["variant_end"].data
+        cn = ds_hmm["call_CN"].data
+        with self._dask_progress(desc="Load CNV HMM data"):
+            pos, end, cn = dask.compute(pos, end, cn)
+
+        debug("access genes")
+        df_genome_features = self.genome_features(region=region)
+        df_genes = df_genome_features.query("type == 'gene'")
+
+        debug("setup intermediates")
+        windows = []
+        modes = []
+        counts = []
+
+        debug("iterate over genes")
+        genes_iterator = self._progress(
+            df_genes.itertuples(),
+            desc="Compute modal gene copy number",
+            total=len(df_genes),
+        )
+        for gene in genes_iterator:
+            # locate windows overlapping the gene
+            loc_gene_start = bisect_left(end, gene.start)
+            loc_gene_stop = bisect_right(pos, gene.end)
+            w = loc_gene_stop - loc_gene_start
+            windows.append(w)
+
+            # slice out copy number data for the given gene
+            cn_gene = cn[loc_gene_start:loc_gene_stop]
+
+            # compute the modes
+            m, c = _cn_mode(cn_gene, vmax=12)
+            modes.append(m)
+            counts.append(c)
+
+        debug("combine results")
+        windows = np.array(windows)
+        modes = np.vstack(modes)
+        counts = np.vstack(counts)
+
+        debug("build dataset")
+        ds_out = xr.Dataset(
+            coords={
+                "gene_id": (["genes"], df_genes["ID"].values),
+                "sample_id": (["samples"], ds_hmm["sample_id"].values),
+            },
+            data_vars={
+                "gene_contig": (["genes"], df_genes["contig"].values),
+                "gene_start": (["genes"], df_genes["start"].values),
+                "gene_end": (["genes"], df_genes["end"].values),
+                "gene_windows": (["genes"], windows),
+                "gene_name": (["genes"], df_genes["Name"].values),
+                "gene_strand": (["genes"], df_genes["strand"].values),
+                "gene_description": (["genes"], df_genes["description"].values),
+                "CN_mode": (["genes", "samples"], modes),
+                "CN_mode_count": (["genes", "samples"], counts),
+                "sample_coverage_variance": (
+                    ["samples"],
+                    ds_hmm["sample_coverage_variance"].values,
+                ),
+                "sample_is_high_variance": (
+                    ["samples"],
+                    ds_hmm["sample_is_high_variance"].values,
+                ),
+            },
+        )
+
+        return ds_out
+
+    def gene_cnv_frequencies(
+        self,
+        region: base_params.regions,
+        cohorts,
+        sample_query=None,
+        min_cohort_size=10,
+        sample_sets=None,
+        drop_invariant=True,
+        max_coverage_variance=DEFAULT_MAX_COVERAGE_VARIANCE,
+    ):
+        """Compute modal copy number by gene, then compute the frequency of
+        amplifications and deletions in one or more cohorts, from HMM data.
+
+        Parameters
+        ----------
+        region: str or list of str or Region or list of Region
+            Chromosome arm (e.g., "2L"), gene name (e.g., "AGAP007280"), genomic
+            region defined with coordinates (e.g., "2L:44989425-44998059") or a
+            named tuple with genomic location `Region(contig, start, end)`.
+            Multiple values can be provided as a list, in which case data will
+            be concatenated, e.g., ["3R", "3L"].
+        cohorts : str or dict
+            If a string, gives the name of a predefined cohort set, e.g., one of
+            {"admin1_month", "admin1_year", "admin2_month", "admin2_year"}.
+            If a dict, should map cohort labels to sample queries, e.g.,
+            `{"bf_2012_col": "country == 'Burkina Faso' and year == 2012 and
+            taxon == 'coluzzii'"}`.
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        min_cohort_size : int
+            Minimum cohort size, below which cohorts are dropped.
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier (e.g., "AG1000G-AO") or a list of
+            sample set identifiers (e.g., ["AG1000G-BF-A", "AG1000G-BF-B"]) or a
+            release identifier (e.g., "3.0") or a list of release identifiers.
+        drop_invariant : bool, optional
+            If True, drop any rows where there is no evidence of variation.
+        max_coverage_variance : float, optional
+            Remove samples if coverage variance exceeds this value.
+
+        Returns
+        -------
+        df : pandas.DataFrame
+            A dataframe of CNV amplification (amp) and deletion (del)
+            frequencies in the specified cohorts, one row per gene and CNV type
+            (amp/del).
+
+        """
+        debug = self._log.debug
+
+        debug("check and normalise parameters")
+        self._check_param_min_cohort_size(min_cohort_size)
+        regions: List[Region] = parse_multi_region(self, region)
+        del region
+
+        debug("access and concatenate data from regions")
+        df = pd.concat(
+            [
+                self._gene_cnv_frequencies(
+                    region=r,
+                    cohorts=cohorts,
+                    sample_query=sample_query,
+                    min_cohort_size=min_cohort_size,
+                    sample_sets=sample_sets,
+                    drop_invariant=drop_invariant,
+                    max_coverage_variance=max_coverage_variance,
+                )
+                for r in regions
+            ],
+            axis=0,
+        )
+
+        debug("add metadata")
+        title = f"Gene CNV frequencies ({region_str(regions)})"
+        df.attrs["title"] = title
+
+        return df
+
+    def _gene_cnv_frequencies(
+        self,
+        *,
+        region,
+        cohorts,
+        sample_query,
+        min_cohort_size,
+        sample_sets,
+        drop_invariant,
+        max_coverage_variance,
+    ):
+        debug = self._log.debug
+
+        debug("sanity check - this function is one region at a time")
+        assert isinstance(region, Region)
+
+        debug("get gene copy number data")
+        ds_cnv = self.gene_cnv(
+            region=region,
+            sample_sets=sample_sets,
+            sample_query=sample_query,
+            max_coverage_variance=max_coverage_variance,
+        )
+
+        debug("load sample metadata")
+        df_samples = self.sample_metadata(sample_sets=sample_sets)
+
+        debug("align sample metadata with samples in CNV data")
+        sample_id = ds_cnv["sample_id"].values
+        df_samples = df_samples.set_index("sample_id").loc[sample_id].reset_index()
+
+        debug("figure out expected copy number")
+        if region.contig == "X":
+            is_male = (df_samples["sex_call"] == "M").values
+            expected_cn = np.where(is_male, 1, 2)[np.newaxis, :]
+        else:
+            expected_cn = 2
+
+        debug(
+            "setup output dataframe - two rows for each gene, one for amplification and one for deletion"
+        )
+        n_genes = ds_cnv.dims["genes"]
+        df_genes = ds_cnv[
+            [
+                "gene_id",
+                "gene_name",
+                "gene_strand",
+                "gene_description",
+                "gene_contig",
+                "gene_start",
+                "gene_end",
+            ]
+        ].to_dataframe()
+        df = pd.concat([df_genes, df_genes], axis=0).reset_index(drop=True)
+        df.rename(
+            columns={
+                "gene_contig": "contig",
+                "gene_start": "start",
+                "gene_end": "end",
+            },
+            inplace=True,
+        )
+
+        debug("add CNV type column")
+        df_cnv_type = pd.DataFrame(
+            {
+                "cnv_type": np.array(
+                    (["amp"] * n_genes) + (["del"] * n_genes), dtype=object
+                )
+            }
+        )
+        df = pd.concat([df, df_cnv_type], axis=1)
+
+        debug("set up intermediates")
+        cn = ds_cnv["CN_mode"].values
+        is_amp = cn > expected_cn
+        is_del = (cn >= 0) & (cn < expected_cn)
+        is_called = cn >= 0
+
+        debug("set up cohort dict")
+        coh_dict = self._locate_cohorts(cohorts=cohorts, df_samples=df_samples)
+
+        debug("compute cohort frequencies")
+        freq_cols = dict()
+        for coh, loc_coh in coh_dict.items():
+            n_samples = np.count_nonzero(loc_coh)
+            debug(f"{coh}, {n_samples} samples")
+
+            if n_samples >= min_cohort_size:
+                # subset data to cohort
+                is_amp_coh = np.compress(loc_coh, is_amp, axis=1)
+                is_del_coh = np.compress(loc_coh, is_del, axis=1)
+                is_called_coh = np.compress(loc_coh, is_called, axis=1)
+
+                # count amplifications and deletions
+                amp_count_coh = np.sum(is_amp_coh, axis=1)
+                del_count_coh = np.sum(is_del_coh, axis=1)
+                called_count_coh = np.sum(is_called_coh, axis=1)
+
+                # compute frequencies, taking accessibility into account
+                with np.errstate(divide="ignore", invalid="ignore"):
+                    amp_freq_coh = np.where(
+                        called_count_coh > 0, amp_count_coh / called_count_coh, np.nan
+                    )
+                    del_freq_coh = np.where(
+                        called_count_coh > 0, del_count_coh / called_count_coh, np.nan
+                    )
+
+                freq_cols[f"frq_{coh}"] = np.concatenate([amp_freq_coh, del_freq_coh])
+
+        debug("build a dataframe with the frequency columns")
+        df_freqs = pd.DataFrame(freq_cols)
+
+        debug("compute max_af and additional columns")
+        df_extras = pd.DataFrame(
+            {
+                "max_af": df_freqs.max(axis=1),
+                "windows": np.concatenate(
+                    [ds_cnv["gene_windows"].values, ds_cnv["gene_windows"].values]
+                ),
+            }
+        )
+
+        debug("build the final dataframe")
+        df.reset_index(drop=True, inplace=True)
+        df = pd.concat([df, df_freqs, df_extras], axis=1)
+        df.sort_values(["contig", "start", "cnv_type"], inplace=True)
+        df.reset_index(drop=True, inplace=True)
+
+        debug("add label")
+        df["label"] = self._pandas_apply(
+            self._make_gene_cnv_label, df, columns=["gene_id", "gene_name", "cnv_type"]
+        )
+
+        debug("deal with invariants")
+        if drop_invariant:
+            df = df.query("max_af > 0")
+
+        debug("set index for convenience")
+        df.set_index(["gene_id", "gene_name", "cnv_type"], inplace=True)
+
+        return df
+
+    def gene_cnv_frequencies_advanced(
+        self,
+        region: base_params.regions,
+        area_by,
+        period_by,
+        sample_sets=None,
+        sample_query=None,
+        min_cohort_size=10,
+        variant_query=None,
+        drop_invariant=True,
+        max_coverage_variance=DEFAULT_MAX_COVERAGE_VARIANCE,
+        ci_method="wilson",
+    ):
+        """Group samples by taxon, area (space) and period (time), then compute
+        gene CNV counts and frequencies.
+
+        Parameters
+        ----------
+        region: str or list of str or Region or list of Region
+            Chromosome arm (e.g., "2L"), gene name (e.g., "AGAP007280"), genomic
+            region defined with coordinates (e.g., "2L:44989425-44998059") or a
+            named tuple with genomic location `Region(contig, start, end)`.
+            Multiple values can be provided as a list, in which case data will
+            be concatenated, e.g., ["3R", "3L"].
+        area_by : str
+            Column name in the sample metadata to use to group samples spatially.
+            E.g., use "admin1_iso" or "admin1_name" to group by level 1
+            administrative divisions, or use "admin2_name" to group by level 2
+            administrative divisions.
+        period_by : {"year", "quarter", "month"}
+            Length of time to group samples temporally.
+        sample_sets : str or list of str, optional
+            Can be a sample set identifier (e.g., "AG1000G-AO") or a list of
+            sample set identifiers (e.g., ["AG1000G-BF-A", "AG1000G-BF-B"]) or a
+            release identifier (e.g., "3.0") or a list of release identifiers.
+        sample_query : str, optional
+            A pandas query string which will be evaluated against the sample
+            metadata e.g., "taxon == 'coluzzii' and country == 'Burkina Faso'".
+        min_cohort_size : int, optional
+            Minimum cohort size. Any cohorts below this size are omitted.
+        variant_query : str, optional
+        drop_invariant : bool, optional
+            If True, drop any rows where there is no evidence of variation.
+        max_coverage_variance : float, optional
+            Remove samples if coverage variance exceeds this value.
+        ci_method : {"normal", "agresti_coull", "beta", "wilson", "binom_test"}, optional
+            Method to use for computing confidence intervals, passed through to
+            `statsmodels.stats.proportion.proportion_confint`.
+
+        Returns
+        -------
+        ds : xarray.Dataset
+            The resulting dataset contains data has dimensions "cohorts" and
+            "variants". Variables prefixed with "cohort" are 1-dimensional
+            arrays with data about the cohorts, such as the area, period, taxon
+            and cohort size. Variables prefixed with "variant" are 1-dimensional
+            arrays with data about the variants, such as the contig, position,
+            reference and alternate alleles. Variables prefixed with "event" are
+            2-dimensional arrays with the allele counts and frequency
+            calculations.
+
+        """
+
+        self._check_param_min_cohort_size(min_cohort_size)
+
+        regions: List[Region] = parse_multi_region(self, region)
+        del region
+
+        ds = simple_xarray_concat(
+            [
+                self._gene_cnv_frequencies_advanced(
+                    region=r,
+                    area_by=area_by,
+                    period_by=period_by,
+                    sample_sets=sample_sets,
+                    sample_query=sample_query,
+                    min_cohort_size=min_cohort_size,
+                    variant_query=variant_query,
+                    drop_invariant=drop_invariant,
+                    max_coverage_variance=max_coverage_variance,
+                    ci_method=ci_method,
+                )
+                for r in regions
+            ],
+            dim="variants",
+        )
+
+        title = f"Gene CNV frequencies ({region_str(regions)})"
+        ds.attrs["title"] = title
+
+        return ds
+
+    def _gene_cnv_frequencies_advanced(
+        self,
+        *,
+        region,
+        area_by,
+        period_by,
+        sample_sets,
+        sample_query,
+        min_cohort_size,
+        variant_query,
+        drop_invariant,
+        max_coverage_variance,
+        ci_method,
+    ):
+        debug = self._log.debug
+
+        debug("sanity check - here we deal with one region only")
+        assert isinstance(region, Region)
+
+        debug("access gene CNV calls")
+        ds_cnv = self.gene_cnv(
+            region=region,
+            sample_sets=sample_sets,
+            sample_query=sample_query,
+            max_coverage_variance=max_coverage_variance,
+        )
+
+        debug("load sample metadata")
+        df_samples = self.sample_metadata(sample_sets=sample_sets)
+
+        debug("align sample metadata")
+        sample_id = ds_cnv["sample_id"].values
+        df_samples = df_samples.set_index("sample_id").loc[sample_id].reset_index()
+
+        debug("prepare sample metadata for cohort grouping")
+        df_samples = self._prep_samples_for_cohort_grouping(
+            df_samples=df_samples,
+            area_by=area_by,
+            period_by=period_by,
+        )
+
+        debug("group samples to make cohorts")
+        group_samples_by_cohort = df_samples.groupby(["taxon", "area", "period"])
+
+        debug("build cohorts dataframe")
+        df_cohorts = self._build_cohorts_from_sample_grouping(
+            group_samples_by_cohort, min_cohort_size
+        )
+
+        debug("figure out expected copy number")
+        if region.contig == "X":
+            is_male = (df_samples["sex_call"] == "M").values
+            expected_cn = np.where(is_male, 1, 2)[np.newaxis, :]
+        else:
+            expected_cn = 2
+
+        debug("set up intermediates")
+        cn = ds_cnv["CN_mode"].values
+        is_amp = cn > expected_cn
+        is_del = (cn >= 0) & (cn < expected_cn)
+        is_called = cn >= 0
+
+        debug("set up main event variables")
+        n_genes = ds_cnv.dims["genes"]
+        n_variants, n_cohorts = n_genes * 2, len(df_cohorts)
+        count = np.zeros((n_variants, n_cohorts), dtype=int)
+        nobs = np.zeros((n_variants, n_cohorts), dtype=int)
+
+        debug("build event count and nobs for each cohort")
+        for cohort_index, cohort in enumerate(df_cohorts.itertuples()):
+            # construct grouping key
+            cohort_key = cohort.taxon, cohort.area, cohort.period
+
+            # obtain sample indices for cohort
+            sample_indices = group_samples_by_cohort.indices[cohort_key]
+
+            # select genotype data for cohort
+            cohort_is_amp = np.take(is_amp, sample_indices, axis=1)
+            cohort_is_del = np.take(is_del, sample_indices, axis=1)
+            cohort_is_called = np.take(is_called, sample_indices, axis=1)
+
+            # compute cohort allele counts
+            np.sum(cohort_is_amp, axis=1, out=count[::2, cohort_index])
+            np.sum(cohort_is_del, axis=1, out=count[1::2, cohort_index])
+
+            # compute cohort allele numbers
+            cohort_n_called = np.sum(cohort_is_called, axis=1)
+            nobs[:, cohort_index] = np.repeat(cohort_n_called, 2)
+
+        debug("compute frequency")
+        with np.errstate(divide="ignore", invalid="ignore"):
+            # ignore division warnings
+            frequency = np.where(nobs > 0, count / nobs, np.nan)
+
+        debug("make dataframe of variants")
+        with warnings.catch_warnings():
+            # ignore "All-NaN slice encountered" warnings
+            warnings.simplefilter("ignore", category=RuntimeWarning)
+            max_af = np.nanmax(frequency, axis=1)
+        df_variants = pd.DataFrame(
+            {
+                "contig": region.contig,
+                "start": np.repeat(ds_cnv["gene_start"].values, 2),
+                "end": np.repeat(ds_cnv["gene_end"].values, 2),
+                "windows": np.repeat(ds_cnv["gene_windows"].values, 2),
+                # alternate amplification and deletion
+                "cnv_type": np.tile(np.array(["amp", "del"]), n_genes),
+                "max_af": max_af,
+                "gene_id": np.repeat(ds_cnv["gene_id"].values, 2),
+                "gene_name": np.repeat(ds_cnv["gene_name"].values, 2),
+                "gene_strand": np.repeat(ds_cnv["gene_strand"].values, 2),
+            }
+        )
+
+        debug("add variant label")
+        df_variants["label"] = self._pandas_apply(
+            self._make_gene_cnv_label,
+            df_variants,
+            columns=["gene_id", "gene_name", "cnv_type"],
+        )
+
+        debug("build the output dataset")
+        ds_out = xr.Dataset()
+
+        debug("cohort variables")
+        for coh_col in df_cohorts.columns:
+            ds_out[f"cohort_{coh_col}"] = "cohorts", df_cohorts[coh_col]
+
+        debug("variant variables")
+        for snp_col in df_variants.columns:
+            ds_out[f"variant_{snp_col}"] = "variants", df_variants[snp_col]
+
+        debug("event variables")
+        ds_out["event_count"] = ("variants", "cohorts"), count
+        ds_out["event_nobs"] = ("variants", "cohorts"), nobs
+        ds_out["event_frequency"] = ("variants", "cohorts"), frequency
+
+        debug("deal with invariants")
+        if drop_invariant:
+            loc_variant = df_variants["max_af"].values > 0
+            ds_out = ds_out.isel(variants=loc_variant)
+            df_variants = df_variants.loc[loc_variant].reset_index(drop=True)
+
+        debug("apply variant query")
+        if variant_query is not None:
+            loc_variants = df_variants.eval(variant_query).values
+            ds_out = ds_out.isel(variants=loc_variants)
+
+        debug("add confidence intervals")
+        self._add_frequency_ci(ds_out, ci_method)
+
+        debug("tidy up display by sorting variables")
+        ds_out = ds_out[sorted(ds_out)]
+
+        return ds_out
 
     def _block_jackknife_cohort_diversity_stats(
         self, *, cohort_label, ac, n_jack, confidence_level
@@ -6044,3 +6656,48 @@ def _moving_h1x(ha, hb, size, start=0, stop=None, step=None):
     out = np.array([_h1x(ha[i:j], hb[i:j]) for i, j in windows])
 
     return out
+
+
+@numba.njit("Tuple((int8, int64))(int8[:], int8)")
+def _cn_mode_1d(a, vmax):
+    # setup intermediates
+    m = a.shape[0]
+    counts = np.zeros(vmax + 1, dtype=numba.int64)
+
+    # initialise return values
+    mode = numba.int8(-1)
+    mode_count = numba.int64(0)
+
+    # iterate over array values, keeping track of counts
+    for i in range(m):
+        v = a[i]
+        if 0 <= v <= vmax:
+            c = counts[v]
+            c += 1
+            counts[v] = c
+            if c > mode_count:
+                mode = v
+                mode_count = c
+            elif c == mode_count and v < mode:
+                # consistency with scipy.stats, break ties by taking lower value
+                mode = v
+
+    return mode, mode_count
+
+
+@numba.njit("Tuple((int8[:], int64[:]))(int8[:, :], int8)")
+def _cn_mode(a, vmax):
+    # setup intermediates
+    n = a.shape[1]
+
+    # setup outputs
+    modes = np.zeros(n, dtype=numba.int8)
+    counts = np.zeros(n, dtype=numba.int64)
+
+    # iterate over columns, computing modes
+    for j in range(a.shape[1]):
+        mode, count = _cn_mode_1d(a[:, j], vmax)
+        modes[j] = mode
+        counts[j] = count
+
+    return modes, counts
